pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lol99"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.1",
    "black>=25.1.0",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "ruff>=0.12.7",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
    "aiomysql>=0.2.0",
    "isort>=5.13.2",
    "cryptography>=45.0.5",
    "types-requests>=2.32.4.20250611",
    "asyncpg>=0.30.0",
    "python-jose>=3.3.0",
    "python-dotenv>=1.0.0",
    "python-jose>=3.5.0",
    "itsdangerous>=2.2.0",
]

[tool.setuptools]
packages = ["app"]
[tool.aerich]
tortoise_orm = "app.config.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[tool.ruff]
line-length = 88
fix = true
exclude = [
    "migrations/",
    "tests/", # 테스트 디렉토리도 보통 제외합니다.
    "venv/",
    ".venv/",
    "__pypackages__/",
]

[tool.ruff.lint]
select = ["I", "E", "F", "W"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]
exclude = ["build", "migrations", "venv", ".venv"]

[[tool.mypy.overrides]]
module = "app.*"
follow_imports = "normal"

